Reorganize files, disable babylon stuff for now
