+++ package.json
@@ -12,6 +12,7 @@
     "leva": "^0.9.34",
     "matter-attractors-f": "^0.1.7",
     "matter-js": "^0.19.0",
+    "r3f-perf": "^7.1.2",
     "react": "^18.2.0",
     "react-burger-menu": "^3.0.9",
     "react-dom": "^18.2.0",
