+++ src/components/Babylon/Base.jsx
@@ -1,11 +1,8 @@
 import * as BABYLON from "babylonjs";
 import { Component } from "react";
-import Stats from "../../../libs/Stats";
 
 export default class Base extends Component {
   componentDidMount() {
-    Stats.mount();
-
     this.renderer = document.createElement("canvas");
     this.renderer.style = "width:100%; height:100%;";
     document.body.appendChild(this.renderer);
@@ -14,21 +11,11 @@ export default class Base extends Component {
   }
 
   componentWillUnmount() {
-    Stats.unmount();
     this.renderer.remove();
     this.engine.stopRenderLoop();
   }
 
-  beforeRender() {
-    Stats.begin();
-  }
+  beforeRender() {}
 
-  afterRender() {
-    Stats.end();
-  }
-
-  // Provide interface to dat.gui;
-  gui() {
-    // return GUI.interface();
-  }
+  afterRender() {}
 }
