+++ src/components/Babylon/ShaderLayering/atmosphereFragment.glsl
