+++ src/components/Babylon/ShaderLayering/vertex.glsl
