+++ src/components/Javascript/index.js
@@ -1,6 +1,8 @@
-import MatterRepulsor from "./Experiments/MatterRepulsor";
+import MatterRepulsor from "./MatterRepulsor";
+import Typography from "./Typography";
 
 const Javascript = {
+  Typography,
   MatterRepulsor,
 };
