--- src/components/Three/Experiments/Base.jsx
@@ -1,25 +1,0 @@
-import { Component } from "react";
-import Stats from "../../../libs/Stats";
-
-export default class Base extends Component {
-  constructor({}) {
-    super();
-  }
-
-  componentDidMount() {
-    Stats.mount();
-  }
-
-  componentWillUnmount() {
-    Stats.unmount();
-    cancelAnimationFrame(this.raf);
-  }
-
-  beforeRender() {
-    Stats.begin();
-  }
-
-  afterRender() {
-    Stats.end();
-  }
-}
