+++ src/components/Three/assets/particletextureshaded.png
