+++ src/components/Three/components/Ground.jsx
@@ -1,7 +1,7 @@
 import React, { useRef } from "react";
 import { useFrame } from "@react-three/fiber";
 import { Color } from "three";
-import "../../../Materials/PerlinNoiseMaterial";
+import "../../Materials/PerlinNoiseMaterial";
 
 export default function Ground(props) {
   const { source = null, color = new Color(0x111111) } = props;
@@ -12,11 +12,11 @@ export default function Ground(props) {
 
     if (source) {
       const { mental } = source.sample();
-      matRef.current.lineColor.setHSL(0, 0, mental);
+      matRef.current.lineColor.setHSL(0.9 * (10 - mental), 1, mental);
       matRef.current.elevation = 1 * mental;
       matRef.current.time += mental / 4;
-      matRef.current.sombrero_frequency = 0.5 * mental;
-      matRef.current.sombrero_amplitude = 0.3 * mental;
+      matRef.current.sombrero_frequency = 1.5 * mental;
+      matRef.current.sombrero_amplitude = 5.3 * mental;
     } else {
       matRef.current.elevation = Math.sin(matRef.current.time / 4) / 2;
     }
