+++ src/components/Three/sass/audiovisualizer.scss
@@ -6,6 +6,7 @@
   align-items: center;
   z-index: 10000;
   display: flex;
+  opacity: 0;
 
   .player-controls-button {
     color: #fff;
