--- src/libs/Palette/index.js
@@ -1,31 +1,0 @@
-const Personal = {
-  lighter: 0xffffff,
-  light: 0xaaaaaa,
-  normal: 0xc82222,
-  dark: 0x666666,
-  darker: 0x222222,
-};
-
-const Synth = {
-  lighter: 0x33a7cc, // CYAN (85,231,255)
-  light: 0x00ccfd, // BLUE (0,204,253)
-  normal: 0xff34b3, // PURPLE (255,52,179)
-  dark: 0x2011a2, // DARKBLUE (32,17,162)
-  darker: 0x201148, // DEEPBLUE (32,17,72)
-};
-
-const Lava = {
-  lighter: 0xeadab5, // (234,218,181)
-  light: 0xffff44, // (255,255,68)
-  normal: 0xff6600, // (255,102,0)
-  dark: 0xcc4422, // (204,68,34)
-  darker: 0x553333, // (85,51,51)
-};
-
-const Palette = {
-  Personal,
-  Synth,
-  Lava,
-};
-
-export default Palette;
