+++ src/libs/palette.js
@@ -1,0 +1,31 @@
+const Personal = {
+  lighter: "#ffffff",
+  light: "#aaaaaa",
+  normal: "#c82222",
+  dark: "#666666",
+  darker: "#222222",
+};
+
+const Synth = {
+  lighter: "#33a7cc", // CYAN (85,231,255)
+  light: "#00ccfd", // BLUE (0,204,253)
+  normal: "#ff34b3", // PURPLE (255,52,179)
+  dark: "#2011a2", // DARKBLUE (32,17,162)
+  darker: "#201148", // DEEPBLUE (32,17,72)
+};
+
+const Lava = {
+  lighter: "#eadab5", // (234,218,181)
+  light: "#ffff44", // (255,255,68)
+  normal: "#ff6600", // (255,102,0)
+  dark: "#cc4422", // (204,68,34)
+  darker: "#553333", // (85,51,51)
+};
+
+const Palette = {
+  Personal,
+  Synth,
+  Lava,
+};
+
+export default Palette;
