+++ src/urls.js
@@ -72,11 +72,11 @@ const urls = [
     children: three,
     icon: () => <SiThreedotjs size="18" />,
   },
-  {
-    name: "Babylon",
-    children: babylon,
-    icon: () => <HiOutlineCubeTransparent size="18" />,
-  },
+  // {
+  //   name: "Babylon",
+  //   children: babylon,
+  //   icon: () => <HiOutlineCubeTransparent size="18" />,
+  // },
   {
     name: "Javascript",
     children: javascript,
